Header/Footer
I made a header that has a logo on the top left that brings you to the landing page. It also has some navigation links that work with placeholder content. On the top right there's a login and signup button. My footer is very simple and just has a copyright text.