Show details Pop up
I had the idea to create a pop up like in Netflix where you click the show and you see some of the details and watch the show. So, I made it so if you click the show card, you can see some more details and a description of the show along with the watched and watchlist button.  This functionality is on all pages except for the landing page.