trying to display addList from database and added login and signup api endpoints.
I added login and signup endpoints where the signup endpoint creates a new record in my users table. The login one checks for the username and password then creates a jwt token.

For the addList I manually added some ids to the add field for a user. I then made an endpoint to give those ids to the frontend. After that I figured out how to display them.