Rating System
I added a frontend, so when you get the show details for a watched show you can add a rating.

I also added the ability to add ratings to the database and get all the ratings for a user when you login.

Lastly I added the rating to the recommendation system, so if you're doing recommendations based on your watched shows, you weight the semantic score based on the rating of the show you get the semantic score from.