Sidebar component and reworking recommendation UI
I added a sidebar, so you can access functionalities of the recommendation system through the sidebar.

For the recommendations I cleaned up the UI so it isn't as confusing. For example instead of having to search for a show by its title, you can use filters and search then click a button to add it to a list.

I also changed the filtering system, so you don't have to make an api call every time you filter the recommended  shows.